scripts/cert_sync/requirements.txt
kink==0.8.1
oci-cli==3.54.5
doppler-sdk==1.3.0
slack_sdk==3.35.0
